Fix 7 bugs found during second code review
- contest.cpp: fix integer division in setTqrm() that prevented QRM
  from ever triggering (512/11025/240 always truncated to 0)
- contest.cpp: add mutex lock in time() for thread safety
- random.cpp: clamp integers() return to prevent off-by-one when
  random() returns exactly 1.0
- mystation.cpp: add envelope bounds check in updateCallInMessage()
- keyer.cpp: replace float accumulation loop with integer loop in
  setRisetime() to avoid drift
- qrmstation.cpp: add explicit cast to fix signed/unsigned mismatch
- station.cpp: guard timeout decrement with NEVER check in tick()

+++ qrmstation.cpp
@@ -39,7 +39,7 @@ QrmStation::QrmStation(RNG *rng, Keyer *keyer, CallList *callList,
 	this->wpm = rng->integers(30, 51);
 
 	// Send a random QRM message to start
-	int msg_idx = rng->integers(0, qrm_messages.size());
+	int msg_idx = rng->integers(0, static_cast<int>(qrm_messages.size()));
 	sendMsg(qrm_messages[msg_idx]);
 }
